pyproject.toml
[project]
name = "cmd-line-executor"
version = "0.1.0"
description = "Example MCP server to call command line apps"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
 "mcp>=1.0.0",
 "python-dotenv>=1.0.1",
]
[[project.authors]]
name = "Markus Pfundstein"
email = "markus@life-electronic.nl"

[build-system]
requires = [ "hatchling",]
build-backend = "hatchling.build"

[project.scripts]
cmd-line-executor = "cmd_line_executor:main"
